revamped the spawn menu

+++ Base/Source/Main/Game/Systems/BattleSystem.cpp
@@ -131,7 +131,7 @@ void BattleSystem::UpdateCharacterLogic(std::vector<CharacterEntity*>& Character
 			Vector3 Direction = it->GetVelocity();
 			if (it->GetVelocity().LengthSquared() > it->WalkSpeed * it->WalkSpeed)
 			{
-				it->SetVelocity(Direction.Normalize() * it->WalkSpeed);
+				it->SetVelocity(Direction.Normalize() * (float)it->WalkSpeed);
 			}
 		}
 		it->Update(dt);
